feat(cuda): float16 and float64 value dtypes
- add f16/f64 to the fused tp + conv kernels, drop the int32 value path
- ops inherit the operand dtype: `_fused_eval` promotes x/y and packs the
  coefficients in it, so the buffers and the coef stride always agree
- f16 not on atomicAdd paths (LEADING_CHANNELS tp, scatter_add_1): raise at trace
- f64 needs jax_enable_x64, which also flips the default int dtype, so GraphCSR
  derives its CSR buffers as int32 explicitly
- SparseMixin aggregates in the summand dtype rather than the x64 default
- copy_pipe sizes by bytes now: old code over-copied at N>1 (f64 N=4 overran the
  buffer), f32 conv same bytes but better coalesced

BREAKING CHANGE: int32 value dtype dropped from the fused kernels. scatter_add_1
with int32 values returns InvalidArgument. index dtypes (int32/uint8) unchanged.

+++ src/e3j/ops/convolution.py
@@ -156,7 +156,7 @@ def convolution(
         if y_parity is None:
             raise ValueError("SENDER graph ordering requires `y_parity`.")
         with jax.ensure_compile_time_eval():
-            c = Coef4D.unpack(coef, val_dtype="float32")
+            c = Coef4D.unpack(coef, val_dtype=x.dtype)
             signs = jnp.asarray(y_parity, dtype=c.val.dtype)[c.idx[:, 2]]
             coef_fwd = Coef4D(
                 c.val * signs, c.idx, val_dtype=c.val_dtype, idx_dtype=c.idx_dtype
@@ -318,7 +318,7 @@ def convolution_bwd(
         #   - dx = bigotimes(coef_dx, dm, y, s)
         #   - dy = bigotimes(coef_dy, dm, x, s)
         #   - ds = bigotimes(coef_ds, dm, y, x)
-        c = Coef4D.unpack(coef, val_dtype="float32")
+        c = Coef4D.unpack(coef, val_dtype=x.dtype)
         coef_dx = c.transpose((1, 0, 2, 3)).pack_jax()
         coef_dy = c.transpose((2, 0, 1, 3)).pack_jax()
         coef_ds = c.transpose((3, 0, 2, 1)).pack_jax()
